+++ src/main/java/com/keunsori/keunsoriserver/domain/reservation/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package com.keunsori.keunsoriserver.reservation.controller;
+package com.keunsori.keunsoriserver.domain.reservation.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -11,10 +11,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.keunsori.keunsoriserver.service.ReservationService;
-import com.keunsori.keunsoriserver.reservation.dto.ReservationCreateRequest;
-import com.keunsori.keunsoriserver.reservation.dto.ReservationResponse;
-import com.keunsori.keunsoriserver.reservation.dto.ReservationUpdateRequest;
+import com.keunsori.keunsoriserver.domain.reservation.dto.ReservationUpdateRequest;
+import com.keunsori.keunsoriserver.domain.reservation.service.ReservationService;
+import com.keunsori.keunsoriserver.domain.reservation.dto.ReservationCreateRequest;
+import com.keunsori.keunsoriserver.domain.reservation.dto.ReservationResponse;
 
 import java.net.URI;
 import java.util.List;
