+++ src/main/java/com/keunsori/keunsoriserver/domain/reservation/domain/vo/ReservationType.java
@@ -1,0 +1,5 @@
+package com.keunsori.keunsoriserver.domain.reservation.domain.vo;
+
+public enum ReservationType {
+    TEAM, PERSONAL
+}
