+++ src/main/java/com/keunsori/keunsoriserver/domain/reservation/domain/vo/Session.java
@@ -1,4 +1,4 @@
-package com.keunsori.keunsoriserver.reservation.domain.vo;
+package com.keunsori.keunsoriserver.domain.reservation.domain.vo;
 
 public enum Session {
     VOCAL, KEYBOARD, DRUM, GUITAR, BASS, ALL
