+++ src/main/java/com/keunsori/keunsoriserver/domain/reservation/repository/ReservationRepository.java
@@ -1,9 +1,9 @@
-package com.keunsori.keunsoriserver.reservation.repository;
+package com.keunsori.keunsoriserver.domain.reservation.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import com.keunsori.keunsoriserver.domain.reservation.domain.Reservation;
 import com.keunsori.keunsoriserver.member.Member;
-import com.keunsori.keunsoriserver.reservation.domain.Reservation;
 
 import java.util.List;
