+++ src/main/java/com/keunsori/keunsoriserver/domain/reservation/service/ReservationService.java
@@ -1,14 +1,14 @@
-package com.keunsori.keunsoriserver.service;
+package com.keunsori.keunsoriserver.domain.reservation.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.keunsori.keunsoriserver.domain.reservation.domain.Reservation;
+import com.keunsori.keunsoriserver.domain.reservation.dto.ReservationCreateRequest;
+import com.keunsori.keunsoriserver.domain.reservation.dto.ReservationResponse;
+import com.keunsori.keunsoriserver.domain.reservation.dto.ReservationUpdateRequest;
 import com.keunsori.keunsoriserver.member.Member;
-import com.keunsori.keunsoriserver.reservation.domain.Reservation;
-import com.keunsori.keunsoriserver.reservation.dto.ReservationCreateRequest;
-import com.keunsori.keunsoriserver.reservation.dto.ReservationResponse;
-import com.keunsori.keunsoriserver.reservation.dto.ReservationUpdateRequest;
-import com.keunsori.keunsoriserver.reservation.repository.ReservationRepository;
+import com.keunsori.keunsoriserver.domain.reservation.repository.ReservationRepository;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
