--- src/main/java/com/keunsori/keunsoriserver/reservation/domain/vo/ReservationType.java
@@ -1,5 +1,0 @@
-package com.keunsori.keunsoriserver.reservation.domain.vo;
-
-public enum ReservationType {
-    TEAM, PERSONAL
-}
